Do not forcibly re-include config file in retestconfig.php
The default config file will clobber the existing global variables
back to "", and then try to "require_once($db_file);", which will
do nothing because it was already included. Then it will think the
db config file has not been edited, and output the first-run message
instead of setting up additional variables. This results in
retestconfig.php displaying the initial welcome message instead of a
table of configuration variables

If retestconfig.php wants to re-eval the config file, it will need to
do it in some scope that protects the globals and lets the main config file
re-include the db config file.